+++ changelog.md
@@ -3,7 +3,7 @@
 *4.??.?? (2017/??/??)*
 
 * Fixed issues with setting media dimensions properly on hidden iframe @rafa8626
-* Fixed way to detect fullscreen mode properly by changing `loadstart` by `loadedmetadata` event @rafa8626
+* Fixed way to detect fullscreen mode properly for all renderers and fixed fullscreen for iOS @rafa8626
 
 *4.0.6 (2017/04/20)*